[zk-sdk] Change `FeeSigmaProof` to `PercentageWithCapProof` (#1084)
* rename submodule `fee` with `percentage_with_cap`

* rename `FeeSigmaProof` type to `PercentageWithCapProof`

* rename `FeeMaxProof` and `FeeEqualityProof` with `PercentageMaxProof` and `PercentageEqualityProof`

* remove `fee` from the variable names

* remove `fee` from function names

* rename `FeeSigmaProofVerificationError` to `PercentageWithCapVerificationError`

* flatten out function arguments

* update docs/comments

* update test names

* clippy
