+++ zk-sdk/src/sigma_proofs/mod.rs
@@ -12,8 +12,8 @@ pub mod errors;
 pub mod batched_grouped_ciphertext_validity;
 pub mod ciphertext_ciphertext_equality;
 pub mod ciphertext_commitment_equality;
-pub mod fee;
 pub mod grouped_ciphertext_validity;
+pub mod percentage_with_cap;
 pub mod pubkey;
 pub mod zero_ciphertext;
